requirements.txt
runpod
albumentations
segmentation_models_pytorch
effdet